requirements.txt
pydicom==1.1.0
cassandra-driver==3.14.0